RAG Agent/langgraph-rag-agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-text-splitters>=0.3.0
langsmith>=0.2.0
faiss-cpu>=1.9.0
openai>=1.0.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
